chore(batch-exports): share the S3 integration fixtures across the temporal tests
One test built its own copy of the local object storage integration that the
conftest fixture already provides, and the GCS integration was defined twice
with identical values.

Generated-By: PostHog Desktop
Task-Id: 25c2e0bb-7208-4c1a-bbae-c2c528424ea3
