+++ products/batch_exports/backend/tests/temporal/destinations/s3/test_workflow_with_gcs_bucket.py
@@ -1,12 +1,10 @@
 import os
-import uuid
 
 import pytest
 
 import aioboto3
 import pytest_asyncio
 
-from posthog.models.integration import Integration
 from posthog.temporal.tests.utils.models import acreate_batch_export, adelete_batch_export
 
 from products.batch_exports.backend.service import BatchExportModel, BatchExportSchema
@@ -45,21 +43,6 @@ async def s3_client(bucket_name, s3_key_prefix):
         yield s3_client
 
         await delete_all_from_s3(s3_client, bucket_name, key_prefix=s3_key_prefix)
-
-
-@pytest_asyncio.fixture
-async def gcs_integration(ateam):
-    """An s3-compatible Integration pointing at GCS, using the AWS credentials in the environment."""
-    return await Integration.objects.acreate(
-        team_id=ateam.pk,
-        kind=Integration.IntegrationKind.S3_COMPATIBLE,
-        integration_id=f"gcs-{uuid.uuid4()}",
-        config={"name": "gcs-test", "endpoint_url": "https://storage.googleapis.com"},
-        sensitive_config={
-            "aws_access_key_id": os.getenv("AWS_ACCESS_KEY_ID"),
-            "aws_secret_access_key": os.getenv("AWS_SECRET_ACCESS_KEY"),
-        },
-    )
 
 
 @pytest_asyncio.fixture
